working basic version

+++ pwrAPI/dummy.c
@@ -7,7 +7,7 @@ int main() {
 	Object obj;
 	obj.type = PWR_OBJ_NODE;
 	PWR_Time now;
-	uint64_t value = 0;
+	double value = 0;
 	PWR_AttrName type = PWR_ATTR_POWER;
 	int ret = PWR_ObjAttrGetValue((void *)&obj, type, (void *)&value, &now);
 	printf("The node power is %lf Watts and the time is %ld\n", value, now);
@@ -16,4 +16,3 @@ int main() {
 	return 0;
 
 }
-
